// src/app/Models/requestItem.ts
export interface RequestItem{
    item_id: number;
    request_id?: number;
    quantity?: number;
    other_details?: string;
    item_unit_id?: number;
    item_unit_name?: string;
    item_list_id?: number;
    item_list_name?: string;
    item_type_id: number;
    item_type_name?: string;
    item_name?: string;
    hayaa_responce?: number;
    response_name?: number;
  }
  